Explicitly define attribute function for LayerFeaturePermutation (#1274)
Summary:

Currently, we inherit attribute function from `LayerFeatureAblation` for `LayerFeaturePermutation`. This diff change it to a standalone version since the signature is different.

Note: There are quite some duplicate logic in `LayerFeatureAblation.attribute` and `LayerFeaturePermutation.attribute`, tried to reuse them but no luck. Open to suggestion on any potential improvement!

Differential Revision: D56275760