Explicitly define attribute function for LayerFeaturePermutation (#1274)
Summary:
Pull Request resolved: https://github.com/pytorch/captum/pull/1274

Currently, we inherit attribute function from `LayerFeatureAblation` for `LayerFeaturePermutation`. This diff change it to a standalone version since the signature is different.

Note: There are quite some duplicate logic in `LayerFeatureAblation.attribute` and `LayerFeaturePermutation.attribute`, tried to reuse them but no luck. Open to suggestion on any potential improvement!

Reviewed By: vivekmig

Differential Revision: D56275760

fbshipit-source-id: 940fe1f917cc2abb04f522622ad109b90894a650