Separate out future logic from attribute method into attribute_future in Feature Ablation (#1320)
Summary:
Pull Request resolved: https://github.com/pytorch/captum/pull/1320

This diff separates the logic for futures in the FeatureAblation class into its own method, attribute_future. This method should only be used with forward functions that return a future, and additional test cases were added to ensure the proper usage of the attribute and the attribute_future methods.

the construct_future_forward helper method was also moved into the utils class to share with other method tests

Reviewed By: yucu

Differential Revision: D60911353

fbshipit-source-id: 516e21e7268c933a0540d6a8f4186a04d189218a