Fix pyre error in _generate_baseline_single_dict_feature
Summary:
Fix the pyre error in `_generate_baseline_single_dict_feature`

The issue was that the device parameter didn't have a type annotation. It appears that the correct type is torch.device.

Also changed the call to `.to()` by adding the `device=` named parameter. If a named parameter is not used in this case, torch may assume this is a dtype (see: https://pytorch.org/docs/stable/generated/torch.Tensor.to.html). Adding the named parameter also clarifies the intent.

Reviewed By: jjuncho

Differential Revision: D65914877