Fix Lime output dimension in batch forward (#1513)
Summary:
Pull Request resolved: https://github.com/pytorch/captum/pull/1513

Currently, when a batch of inputs is provided with a forward function that returns a single scalar per batch, Lime and KernelShap still return output matching the input shape.

This behavior is inconsistent with other perturbation based methods, particularly Feature Ablation and Shapley Value Sampling.

This change breaks backward compatibility for OSS users, but since it's a specific case (scalar per batch), should be fine to update with only a documentation update.

Reviewed By: craymichael

Differential Revision: D70096644

fbshipit-source-id: de9207813240cb4a013680854a363df70ee777f9