Adding test for output that is a tensor of integers. Updating passthrough layer. (#1526)
Summary:
Pull Request resolved: https://github.com/pytorch/captum/pull/1526

We are adding tests for different types of unsupported and non-differentiable layer output. Here, we add a test for layer output that is a tensor of integers.

We split by the cases for unsupported layers from the case when the layer output is used by some tasks and not others.

When layer output is not supported (layer output is a List of Tensors or a Tensor of integers), we don't get attributions and return None for those layers.

In the case when a layer output is not used by a task, we should output a tensor of zeros for that task.

Differential Revision: D70919347