Add min_examples_per_batch to FeatureAblation/FeaturePermutation (#1533)
Summary:
Pull Request resolved: https://github.com/pytorch/captum/pull/1533

For sparse features, the 0th dim is not always the batch size.
Currently for FeatureAblation, we skip the input tensor if there are no elements in it. For FeaturePermutation, we throw if we cannot permute across the "batch" (i.e. the tensor has only one element).
`min_examples_per_batch` parameterizes the above behavior. For FeaturePermutation, the default is set to 2, meaning we no longer throw by default, but return an attribution score of 0

Reviewed By: jjuncho

Differential Revision: D71980024

fbshipit-source-id: 89c9ab5bdfcc0c0cc622d53537f664f1c128db23