Skip feature group when permuting features if any have batch size <= 1 (#1539)
Summary:

Also if all of the associated input tensors are empty, for permutation and ablation.

Basically the same as the previous diff, but for `enable_cross_tensor_attribution=True`. Another attribute was introduced because by default, we want to ablate a feature group even if only one tensor is empty.

Reviewed By: cyrjano

Differential Revision: D72253679