Skip feature group when permuting features if any have batch size <= 1 (#1539)
Summary:
Pull Request resolved: https://github.com/pytorch/captum/pull/1539

Also if all of the associated input tensors are empty, for permutation and ablation.

Basically the same as the previous diff, but for `enable_cross_tensor_attribution=True`. Another attribute was introduced because by default, we want to ablate a feature group even if only one tensor is empty.

Reviewed By: cyrjano

Differential Revision: D72253679

fbshipit-source-id: 301f1453823b13f51787c57f05c6e48e4a6717a0