Enable pyre type checking in OSS (#1569)
Summary:
Pull Request resolved: https://github.com/pytorch/captum/pull/1569

This adds pyre type checking to OSS GitHub actions. This reduces the back and forth between OSS contributors and Captum team due to differing OSS and internal type checking. mypy is kept enabled as well for now since it handles some things that pyre doesn't, but this diff is needed as pyre catches things that mypy doesn't.

Differential Revision: D75113906