Fix OSS mypy and pyre and conda-related testing and typing issues (#1573)
Summary:
Pull Request resolved: https://github.com/pytorch/captum/pull/1573

As title. Mypy decided to update again. As did pyre. And somehow rust compiler may be needed so I added it.

Issues in https://github.com/pytorch/captum/actions/runs/15359553880/job/43224685521
hopefully will just disappear, it is due to a wheel not being found and failing to build a wheel as rust compiler isn't installed, but a wheel should definitely be available.

Reviewed By: cyrjano

Differential Revision: D75896598

fbshipit-source-id: 9e526c060cec656ffc295b12169d2d551a833bfb