Starter Task 1: Ensuring TQDM Progress Bar Always Enabled (#1575)
Summary:
Pull Request resolved: https://github.com/pytorch/captum/pull/1575

My task was to remove the redundant progress bar code now that tqdm is a dependency. The problem was that since tqdm is now a dependency, we don't need to make checks for whether or not tqdm is being used, and thus we can remove the simple progress print code in the progress function (as well as the import error try-except code at the very top of the file). I also removed the SimpleProgress class definition, as well as the use_tqdm parameter in the progress function. Since we are no longer using the SimpleProgress object anywhere, I removed it from wherever it was used in the codebase (feature_ablation.py). In the test_progress.py, I removed the tests corresponding to SimpleProgress.

Reviewed By: cyrjano

Differential Revision: D75814260

fbshipit-source-id: 644073ed6a6d0dd0d74f574a98dff5b6b5a45f2b